refactor(Output): decouple output and output components
